+++ src/lib/Scenes/Consignments/Screens/SubmitArtworkOverview/ArtworkDetails/ArtworkDetails.tsx
@@ -4,9 +4,9 @@ import { GlobalStore } from "lib/store/GlobalStore"
 import { BulletedItem, CTAButton, Flex, Spacer, Text } from "palette"
 import React, { useState } from "react"
 import { ErrorView } from "../Components/ErrorView"
-import { createOrUpdateSubmission } from "../utils/createOrUpdateSubmission"
-import { ArtworkDetailsFormModel, artworkDetailsValidationSchema } from "../utils/validation"
 import { ArtworkDetailsForm } from "./ArtworkDetailsForm"
+import { createOrUpdateSubmission } from "./utils/createOrUpdateSubmission"
+import { ArtworkDetailsFormModel, artworkDetailsValidationSchema } from "./validation"
 
 export const ArtworkDetails: React.FC<{ handlePress: () => void }> = ({ handlePress }) => {
   const { submissionId, artworkDetails } = GlobalStore.useAppState(
