+++ src/lib/Scenes/Consignments/Screens/SubmitArtworkOverview/ArtworkDetails/ArtworkDetailsForm.tsx
@@ -3,12 +3,12 @@ import { artworkRarityClassifications } from "lib/utils/artworkRarityClassificat
 import { Box, BulletedItem, Flex, Input, InputTitle, RadioButton, Spacer, Text } from "palette"
 import { Select } from "palette/elements/Select"
 import React, { useState } from "react"
-import { rarityOptions } from "../utils/rarityOptions"
-import { limitedEditionValue } from "../utils/rarityOptions"
-import { ArtworkDetailsFormModel, Location } from "../utils/validation"
 import { ArtistAutosuggest } from "./Components/ArtistAutosuggest"
 import { InfoModal } from "./Components/InfoModal"
 import { LocationAutocomplete } from "./Components/LocationAutocomplete"
+import { rarityOptions } from "./utils/rarityOptions"
+import { limitedEditionValue } from "./utils/rarityOptions"
+import { ArtworkDetailsFormModel, Location } from "./validation"
 
 export const ArtworkDetailsForm: React.FC = () => {
   const { values, setFieldValue } = useFormikContext<ArtworkDetailsFormModel>()
