+++ src/lib/Scenes/Consignments/Screens/SubmitArtworkOverview/ArtworkDetails/utils/createOrUpdateSubmission.ts
@@ -3,9 +3,9 @@ import {
   CreateSubmissionMutationInput,
 } from "__generated__/createConsignmentSubmissionMutation.graphql"
 import { UpdateSubmissionMutationInput } from "__generated__/updateConsignSubmissionMutation.graphql"
-import { createConsignSubmission, updateConsignSubmission } from "../Mutations"
+import { createConsignSubmission, updateConsignSubmission } from "../../Mutations"
+import { ArtworkDetailsFormModel } from "../validation"
 import { limitedEditionValue } from "./rarityOptions"
-import { ArtworkDetailsFormModel } from "./validation"
 
 export type SubmissionInput = CreateSubmissionMutationInput | UpdateSubmissionMutationInput
