+++ src/lib/Scenes/Consignments/Screens/SubmitArtworkOverview/ArtworkDetails/validation.ts
@@ -87,15 +87,3 @@ export const artworkDetailsValidationSchema = Yup.object().shape({
     country: Yup.string(),
   }),
 })
-
-export interface ContactInformationFormModel {
-  userName: string
-  userEmail: string
-  userPhone: string
-}
-
-export const contactInformationValidationSchema = Yup.object().shape({
-  userName: Yup.string().required("Please provide a name").trim(),
-  userEmail: Yup.string().email().required("Please provide a valid Email").trim(),
-  userPhone: Yup.string().required("Please provide a valid phone number").trim(),
-})
