+++ src/lib/Scenes/Consignments/Screens/SubmitArtworkOverview/State/SubmissionModel.tsx
@@ -1,6 +1,9 @@
 import { Action, action } from "easy-peasy"
 import { ConsignmentsSubmissionUtmParams } from "../../../ConsignmentsHome/ConsignmentsSubmissionForm"
-import { artworkDetailsEmptyInitialValues, ArtworkDetailsFormModel } from "../utils/validation"
+import {
+  artworkDetailsEmptyInitialValues,
+  ArtworkDetailsFormModel,
+} from "../ArtworkDetails/validation"
 
 export interface ArtworkSubmissionModel {
   submissionId: string
