+++ src/lib/Scenes/Consignments/Screens/SubmitArtworkOverview/SubmitArtworkOverview.tsx
@@ -6,7 +6,7 @@ import React, { useRef, useState } from "react"
 import { ScrollView } from "react-native"
 import { ArtworkDetails } from "./ArtworkDetails/ArtworkDetails"
 import { ArtworkSubmittedScreen } from "./ArtworkSubmitted"
-import { ContactInformationQueryRenderer } from "./ContactInformation/ContactInformation"
+import { ContactInformation } from "./ContactInformation/ContactInformation"
 import { UploadPhotos } from "./UploadPhotos/UploadPhotos"
 
 interface SubmitArtworkScreenNavigationProps
@@ -44,7 +44,7 @@ export const SubmitArtworkScreen: React.FC<SubmitArtworkScreenNavigationProps> =
       overtitle: "Step 3 of 3",
       title: "Contact Information",
       Content: (
-        <ContactInformationQueryRenderer
+        <ContactInformation
           handlePress={() => {
             navigation.navigate("ArtworkSubmittedScreen")
           }}
